requirements.txt
torch
torchvision
numpy
streamlit
matplotlib
Pillow
opencv-python